refactor: organize clients and models
