+++ aicostmanager/client/__init__.py
@@ -1,0 +1,19 @@
+"""Client package exposing sync and async variants."""
+
+from .exceptions import (
+    AICMError,
+    APIRequestError,
+    MissingConfiguration,
+    UsageLimitExceeded,
+)
+from .sync_client import CostManagerClient
+from .async_client import AsyncCostManagerClient
+
+__all__ = [
+    "AICMError",
+    "APIRequestError",
+    "MissingConfiguration",
+    "UsageLimitExceeded",
+    "CostManagerClient",
+    "AsyncCostManagerClient",
+]
